requirements.txt
Flask
Werkzeug
sqlite3
click
Flask-WTF
Flask-Login
wtforms-components
colour
flaskr
